% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo_engen.R
\name{mo_engen}
\alias{mo_engen}
\title{Monthly Energy Generation Calculator}
\usage{
mo_engen(df, ins_cap, mo_name, net_gen, per_gen)
}
\arguments{
\item{df}{A dataframe}

\item{ins_cap}{numeric column of Installed Capacity}

\item{mo_name}{name of generated numeric column of Monthly Nameplate Generation}

\item{net_gen}{numeric column of Net Generation}

\item{per_gen}{name of generated numeric column of Percent Net Energy Generation}
}
\value{
Dataframe with 2 new columns containing nameplate and percent net generation
}
\description{
This function take the installed capacity of a plant to determines the monthly nameplate generation and determines the net generation as a percent of the nameplate generation as a measure of production
}
\examples{
mo_engen(mon_gen, "Installed.Capacity", "Monthly.Nameplate", "Net.Generation", "Percent.Generation")
}
\keyword{energy}
