% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo_engen_gr.R
\name{mo_engen_gr}
\alias{mo_engen_gr}
\title{Monthly Energy Generation Graphing Function}
\usage{
mo_engen_gr(df, col, plant, para, month, year)
}
\arguments{
\item{df}{A dataframe}

\item{col}{The name column containing plant names (character strings)}

\item{plant}{A character string of the despired plant to subset}

\item{para}{The desired numeric column to be the x axis (net generation, percent, etc)}

\item{month}{The numeric column of months}

\item{year}{The numeric column of years}
}
\value{
Function with return a series of graphs displaying parameter trends overtime by year
}
\description{
This function will generate a series of graphs based on the output from mo_engen() for an energy plant broken up by year displaying parameter trend over time
}
\examples{
mo_engen_gr(mon_gen_mo, "Plant.Name", "Alvoca", "Month", "Year")
}
\keyword{energy}
