% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yr_engen.R
\name{yr_engen}
\alias{yr_engen}
\title{Yearly Energy Generation Summary Function}
\usage{
yr_engen(df, col, plant, year, para)
}
\arguments{
\item{df}{A Data Frame with the calculated monthly energy generation parameters}

\item{col}{The column with the plant names}

\item{plant}{The name of the plant to summarize data for}

\item{year}{The column with years to direct the function what to group the data by before summarizing}

\item{para}{The desired parameter to summarize}
}
\value{
Function will return a summary table for a single energy plant of the mean of a parameter per year
}
\description{
This function will calculate the average value of energy generation for each available year for a single plant.
}
\examples{
yr_engen(mon_gen_mo, "Plant.Name", "Alcova", "Year", "Percent.Generation")
}
\keyword{energy}
