% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yr_engen_gr.R
\name{yr_engen_gr}
\alias{yr_engen_gr}
\title{Trends in Yearly Energy Generation Graphing Function}
\usage{
yr_engen_gr(df, year, para)
}
\arguments{
\item{df}{The dataframe with the summary data to graph}

\item{year}{The column with year values to graph on the X axis}

\item{para}{The column with the parameter values to graph on the y axis}
}
\value{
Function will return a graph for a single energy plant of the mean of a parameter over time
}
\description{
This function will take a summary table generated by yr_engen() and graph the trends over time of the selected parameter for a single plant
}
\examples{
yr_engen_gr(mon_gen_mo_al, "Year", "Percent.Generation")
}
\keyword{energy}
